backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dynaconf>=3.2.12",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pyjwt>=2.10.1",
    "pytest-cov>=7.0.0",
    "pytest>=9.0.1",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
]
